requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.0
sqlparse==0.4.4
typing_extensions==4.10.0
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.0
numpy==1.26.4
sqlparse==0.4.4
typing_extensions==4.10.0
